move track_modifications
